requirements.txt
flask
python-dotenv
requests
pyTelegramBotAPI
playwright
playwright-stealth
